feat: [OFFNAL-83]: api 호출 클라이언트 수정
