+++ src/infrastructure/remote/axios/createBaseAxiosClient.ts
@@ -1,0 +1,7 @@
+import axios from 'axios'
+import { API_URL } from '@env'
+
+export const baseAxiosClient = axios.create({
+  baseURL: API_URL,
+  timeout: 10000,
+})
